Fix wraparound in PROFILER_GET_ELAPSED_ by specifying correct data sizes.

+++ add_profiler_code.py
@@ -219,13 +219,12 @@ void profiler_end(void)
     profiler_h_src = """#pragma once
 
 #include <stdint.h>
-#include "app_macros.h"
 
 #define PROFILER_GET_US() TIM7->CNT
-#define PROFILER_GET_ELAPSED_US(start) PROFILER_GET_US() - start
+#define PROFILER_GET_ELAPSED_US(start) ((uint16_t)PROFILER_GET_US() - (uint16_t)start)
 
 #define PROFILER_GET_MS() TIM2->CNT
-#define PROFILER_GET_ELAPSED_MS(start) PROFILER_GET_MS() - start
+#define PROFILER_GET_ELAPSED_MS(start) ((uint32_t)PROFILER_GET_MS() - (uint32_t)start)
 
 struct profiler_vars
 {
